Update index.html
Clarified line 197 by adding a few commas to provide natural pauses to increase readability. I am new to Backbone and think this is a great resource. I changed:

So we can bind the change listener to individual attributes or if we like simply 'this.on("change", function(model){});' to listen for changes to all attributes of the model.

to 

So we can bind the change listener to individual attributes, or, if we like, simply 'this.on("change", function(model){});' to listen for changes to all attributes of the model.